LayoutTests/imported/w3c/web-platform-tests/css/css-tables/html-to-css-mapping-2-expected.txt

HTML-2-CSS Mapping

Checks that browsers implement properly the html2css stylesheet (rules about borders, disabling border-collapse for clarity)

1: The border of the table element should be 0px hidden gray:
border-width:0px is the initial value;
border-style:hidden comes from table:matches(..., [rules=groups], ...);
border-color:gray comes from table

<table border="0" rules="groups"><tbody><tr><td></td></tr></tbody></table>
2: The border of the table element should be 1px outset black:
border-width:1px comes from table[frame=box] (1/2);
border-style:outset comes from table[frame=box] (2/2);
border-color:black comes from table:matches(..., [rules=groups], ...)

<table border="0" rules="groups" frame="box"><tbody><tr><td></td></tr></tbody></table>
3a: The border of the table element should be 10px outset gray:
border-width:10px comes from table[border=$w];
border-style:outset comes from table[frame=box] (2/2);
border-color:gray comes from table[border=$w]

3b: The border-top of the td element should be 1px inset gray:
border-width:1px comes from table[border=$w] > ...td;
border-style:inset comes from table[border=$w] > ...td;
border-color:gray comes from table[border=$w] > ...td

3c: The border-left of the td element should be 1px hidden gray:
border-width:1px comes from table[border=$w] > ...td;
border-style:inset comes from table[border=$w] > ...td;
border-color:gray comes from table[border=$w] > ...td

<table border="10" rules="groups" frame="box"><tbody><tr><td></td></tr></tbody></table>
4a: The border of the table element should be 1px outset gray:
border-width:10px comes from table[border=$w];
border-style:outset comes from table[border=$w];
border-color:gray comes from table[border=$w]

4b: The top border of the td element should be 1px solid gray:
border-width:1px comes from table[rules=all] > ... td;
border-style:solid comes from table[rules=all] > ... td;
border-color:gray comes from table[rules=all] > ... td

<table border="yes" rules="all"><tbody><tr><td></td></tr></tbody></table>
5a: The left border of the table element should be ? hidden ?:
border-style:hidden comes from table[frame=hsides];

5b: The right border of the table element should be 10px solid gray:
border-width:10px comes from table[border=$w];
border-style:solid comes from [style];
border-color:gray comes from table[border=$w]

<table border="10" frame="hsides" style="border-right-style: solid"><tbody><tr><td></td></tr></tbody></table>

PASS 1: HTML -> CSS Mapping for borders is applied correctly on table markup (table left width)
PASS 1: HTML -> CSS Mapping for borders is applied correctly on table markup (table left style)
PASS 1: HTML -> CSS Mapping for borders is applied correctly on table markup (table left color)
PASS 2: HTML -> CSS Mapping for borders is applied correctly on table markup (table left width)
FAIL 2: HTML -> CSS Mapping for borders is applied correctly on table markup (table left style) assert_equals: expected "outset" but got "solid"
FAIL 2: HTML -> CSS Mapping for borders is applied correctly on table markup (table left color) assert_equals: expected "rgb(0, 0, 0)" but got "rgb(128, 128, 128)"
FAIL 3a: HTML -> CSS Mapping for borders is applied correctly on table markup (table left width) assert_equals: expected "10px" but got "1px"
FAIL 3a: HTML -> CSS Mapping for borders is applied correctly on table markup (table left style) assert_equals: expected "outset" but got "solid"
PASS 3a: HTML -> CSS Mapping for borders is applied correctly on table markup (table left color)
FAIL 3b: HTML -> CSS Mapping for borders is applied correctly on table markup (td top width) assert_equals: expected "1px" but got "0px"
FAIL 3b: HTML -> CSS Mapping for borders is applied correctly on table markup (td top style) assert_equals: expected "inset" but got "none"
FAIL 3b: HTML -> CSS Mapping for borders is applied correctly on table markup (td top color) assert_equals: expected "rgb(128, 128, 128)" but got "rgb(0, 0, 0)"
FAIL 3c: HTML -> CSS Mapping for borders is applied correctly on table markup (td left width) assert_equals: expected "1px" but got "0px"
FAIL 3c: HTML -> CSS Mapping for borders is applied correctly on table markup (td left style) assert_equals: expected "inset" but got "none"
FAIL 3c: HTML -> CSS Mapping for borders is applied correctly on table markup (td left color) assert_equals: expected "rgb(128, 128, 128)" but got "rgb(0, 0, 0)"
PASS 4: HTML -> CSS Mapping for borders is applied correctly on table markup (table left width)
PASS 4: HTML -> CSS Mapping for borders is applied correctly on table markup (table left style)
PASS 4: HTML -> CSS Mapping for borders is applied correctly on table markup (table left color)
PASS 4: HTML -> CSS Mapping for borders is applied correctly on table markup (td left width)
PASS 4: HTML -> CSS Mapping for borders is applied correctly on table markup (td left style)
PASS 4: HTML -> CSS Mapping for borders is applied correctly on table markup (td left color)
PASS 5a: HTML -> CSS Mapping for borders is applied correctly on table markup (table left style)
FAIL 5b: HTML -> CSS Mapping for borders is applied correctly on table markup (table right width) assert_equals: expected "10px" but got "1px"
PASS 5b: HTML -> CSS Mapping for borders is applied correctly on table markup (table right style)
PASS 5b: HTML -> CSS Mapping for borders is applied correctly on table markup (table right color)
